extract preprocessing of Page3Handler into Helper extract Global Values of Page3Handler fix Filling of empty data lines (!)
